+++ src/javadocofflinesearch/formatters/PlainTextFormatter.java
@@ -38,7 +38,7 @@ public class PlainTextFormatter implements Formatter {
     }
 
     @Override
-    public void title(int current, int totalHits, String title) {
+    public void title(int current, long totalHits, String title) {
         out.println(current + "/" + totalHits + ") " + title);
     }
